fixes for instruqt

+++ Makefile
@@ -4,12 +4,12 @@
 setup_env: setup_env1 setup_env2
 
 setup_env1:
-	cp apps/web-api/.env.example apps/web-api/.env
+	[ ! -f apps/web-api/.env ] && cp apps/web-api/.env.example apps/web-api/.env
 
 setup_env2:
-	cp apps/login-service/.env.example apps/login-service/.env
+	[ ! -f apps/login-service/.env ] && cp apps/login-service/.env.example apps/login-service/.env
 
-apps:
+up:
 	docker compose up
 
 down:
